+++ django_enum/__init__.py
@@ -57,7 +57,7 @@ __all__ = [
     'NonStrictSelectMultiple'
 ]
 
-VERSION = (1, 2, 0)
+VERSION = (1, 3, 0)
 
 __title__ = 'Django Enum'
 __version__ = '.'.join(str(i) for i in VERSION)