+++ django_enum/choices.py
@@ -3,12 +3,85 @@ Support for symmetrical property enumeration types derived from Django choice
 types. These choices types are drop in replacements for the Django
 IntegerChoices and TextChoices.
 """
-import enum
+from enum import Enum, IntFlag
+from typing import Any, List, Optional, Tuple, Type
 
 from django.db.models import Choices
 from django.db.models import IntegerChoices as DjangoIntegerChoices
 from django.db.models import TextChoices as DjangoTextChoices
 from django.db.models.enums import ChoicesMeta
+
+
+def choices(enum: Optional[Type[Enum]]) -> List[Tuple[Any, str]]:
+    """
+    Get the Django choices for an enumeration type. If the enum type has a
+    choices attribute, it will be used. Otherwise, the choices will be derived
+    from value, label pairs if the enumeration type has a label attribute, or
+    the name attribute if it does not.
+
+    This is used for compat with enums that do not inherit from Django's
+    Choices type.
+
+    :param enum: The enumeration type
+    :return: A list of (value, label) pairs
+    """
+    return getattr(
+        enum,
+        'choices', [
+            *([(None, enum.__empty__)] if hasattr(enum, '__empty__') else []),
+            *[
+                (
+                    member.value,
+                    getattr(member, 'label', getattr(member, 'name'))
+                )
+                for member in enum
+            ]
+        ]
+    ) if enum else []
+
+
+def names(enum: Optional[Type[Enum]]) -> List[Any]:
+    """
+    Return a list of names to use for the enumeration type. This is used
+    for compat with enums that do not inherit from Django's Choices type.
+
+    :param enum: The enumeration type
+    :return: A list of labels
+    """
+    return getattr(
+        enum,
+        'names', [
+            *(['__empty__'] if hasattr(enum, '__empty__') else []),
+            *[member.name for member in enum]
+        ]
+    ) if enum else []
+
+
+def labels(enum: Optional[Type[Enum]]) -> List[Any]:
+    """
+    Return a list of labels to use for the enumeration type. See choices.
+
+    This is used for compat with enums that do not inherit from Django's
+    Choices type.
+
+    :param enum: The enumeration type
+    :return: A list of labels
+    """
+    return getattr(enum, 'labels', [label for _, label in choices(enum)])
+
+
+def values(enum: Optional[Type[Enum]]) -> List[Any]:
+    """
+    Return a list of the values of an enumeration type.
+
+    This is used for compat with enums that do not inherit from Django's
+    Choices type.
+
+    :param enum: The enumeration type
+    :return: A list of values
+    """
+    return getattr(enum, 'values', [value for value, _ in choices(enum)])
+
 
 try:
     from enum_properties import (
@@ -73,7 +146,7 @@ try:
     class FlagChoices(  # type: ignore
         DecomposeMixin,
         DjangoSymmetricMixin,
-        enum.IntFlag,
+        IntFlag,
         Choices,
         metaclass=DjangoEnumPropertiesMeta
     ):
@@ -86,7 +159,6 @@ try:
 
 
 except (ImportError, ModuleNotFoundError):
-    from enum import Enum
 
     # 3.11 - extend from Enum so base type check does not throw type error
     class MissingEnumProperties(Enum):
@@ -137,7 +209,7 @@ except (ImportError, ModuleNotFoundError):
 
     class FlagChoices(  # type: ignore
         DjangoSymmetricMixin,
-        enum.IntFlag,
+        IntFlag,
         Choices
     ):
         """Raises ImportError on class definition"""