+++ django_enum/tests/djenum/models.py
@@ -9,6 +9,7 @@ from django_enum.tests.djenum.enums import (
     Constants,
     DJIntEnum,
     DJTextEnum,
+    ExternEnum,
     IntEnum,
     PosIntEnum,
     SmallIntEnum,
@@ -31,6 +32,8 @@ class EnumTester(models.Model):
     constant = EnumField(Constants, null=True, default=None, db_index=True, blank=True)
 
     text = EnumField(TextEnum, null=True, default=None, db_index=True, blank=True)
+
+    extern = EnumField(ExternEnum, null=True, default=None, db_index=True, blank=True)
 
     # basic choice fields - used to compare behavior
     int_choice = models.IntegerField(
@@ -113,3 +116,19 @@ class BadDefault(models.Model):
         default=5,
         blank=True
     )
+
+
+class AdminDisplayBug35(models.Model):
+
+    text_enum = EnumField(
+        DJTextEnum,
+        null=True,
+        default=None
+    )
+
+    int_enum = EnumField(
+        DJIntEnum,
+        null=True,
+        default=None
+    )
+